api/requirements.txt
fastapi
uvicorn[standard]

psycopg2-binary
sqlalchemy
alembic

pydantic
python-dotenv

pandas
numpy

httpx
requests

scikit-learn
matplotlib

openai
